Basic Integration with JSON data complete
As of now, data is being retrieved from a local json file and not from
the '54.169.51.25/api/getCards'  url as I'm on a localhost. As the
basics are done, I can now focus on the functionality of the site,
namely viewing a specific number of cards at a time, displaying the
profile picture and so on and so forth.